# docs/swagger.yaml
definitions:
...
  github_com_evrone_go-clean-template_internal_controller_http_v1_request.Translate:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
    required:
    - destination
    - original
    - source
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.ExamCategory:
    enum:
    - NEET_PG
    - NEET_UG
    - JEE
    - UPSC
    type: string
    x-enum-varnames:
    - ExamCategoryNEETPG
    - ExamCategoryNEETUG
    - ExamCategoryJEE
    - ExamCategoryUPSC
...
  github_com_evrone_go-clean-template_internal_entity.PodcastCreateRequest:
    properties:
      audioUrl:
        type: string
      description:
        type: string
      durationSeconds:
        type: integer
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      isActive:
        type: boolean
      subjectId:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      topicId:
        type: string
    required:
    - audioUrl
    - exam
    - title
    type: object
  github_com_evrone_go-clean-template_internal_entity.PodcastEpisode:
    properties:
      audioUrl:
        type: string
      description:
        type: string
      durationSeconds:
        type: integer
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      id:
        type: string
      isActive:
        type: boolean
      subjectId:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      topicId:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.PracticeAnswerRequest:
    properties:
      selectedOption:
        type: integer
      sessionQuestionId:
        type: string
      timeTakenMs:
        type: integer
    required:
    - selectedOption
    - sessionQuestionId
    type: object
  github_com_evrone_go-clean-template_internal_entity.PracticeMode:
    enum:
    - smart
    - custom
    - revision
    - exam
    type: string
    x-enum-varnames:
    - PracticeModeSmart
    - PracticeModeCustom
    - PracticeModeRevision
    - PracticeModeExam
  github_com_evrone_go-clean-template_internal_entity.PracticeSession:
    properties:
      completedAt:
        type: string
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      id:
        type: string
      mode:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeMode'
      startedAt:
        type: string
      status:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSessionStatus'
      totalQuestionsPlanned:
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_entity.PracticeSessionCreateRequest:
    properties:
      difficultyLevels:
        items:
          type: integer
        type: array
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      mode:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeMode'
      numQuestions:
        type: integer
      subjectIds:
        items:
          type: string
        type: array
      timeLimitMinutes:
        type: integer
      topicIds:
        items:
          type: string
        type: array
    required:
    - mode
    type: object
  github_com_evrone_go-clean-template_internal_entity.PracticeSessionDetail:
    properties:
      questions:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSessionQuestion'
        type: array
      session:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSession'
    type: object
  github_com_evrone_go-clean-template_internal_entity.PracticeSessionQuestion:
    properties:
      answeredAt:
        type: string
      id:
        type: string
      isCorrect:
        type: boolean
      question:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Question'
      selectedOption:
        type: integer
      sequenceIndex:
        type: integer
      timeTakenMs:
        type: integer
    type: object
  github_com_evrone_go-clean-template_internal_entity.PracticeSessionStatus:
    enum:
    - in_progress
    - completed
    - abandoned
    type: string
    x-enum-varnames:
    - PracticeStatusInProgress
    - PracticeStatusCompleted
    - PracticeStatusAbandoned
  github_com_evrone_go-clean-template_internal_entity.Question:
    properties:
      choiceType:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.QuestionChoiceType'
      correctOption:
        type: integer
      difficultyLevel:
        type: integer
      exam:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.ExamCategory'
      explanation:
        type: string
      id:
        type: string
      isActive:
        type: boolean
      isClinical:
        type: boolean
      isHighYield:
        type: boolean
      isImageBased:
        type: boolean
      optionA:
        type: string
      optionB:
        type: string
      optionC:
        type: string
      optionD:
        type: string
      questionText:
        type: string
      subjectId:
        type: string
      topicId:
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.QuestionChoiceType:
    enum:
    - single
    - multi
    type: string
    x-enum-varnames:
    - ChoiceTypeSingle
    - ChoiceTypeMulti
...
  github_com_evrone_go-clean-template_internal_entity.RevisionItem:
    properties:
      id:
        type: string
      intervalIndex:
        type: integer
      nextReviewAt:
        type: string
      question:
        $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Question'
      timesReviewed:
        type: integer
    type: object
...
  github_com_evrone_go-clean-template_internal_entity.Translation:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
      translation:
        example: text for translation
        type: string
    type: object
  github_com_evrone_go-clean-template_internal_entity.TranslationHistory:
    properties:
      history:
        items:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Translation'
        type: array
    type: object
...
  internal_controller_http_v1.ErrorResponse:
    properties:
      error:
        example: message
        type: string
    type: object
info:
  contact: {}
paths:
...
  /admin/podcasts:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastEpisode'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: List podcasts
      tags:
      - 'Admin: Podcasts'
    post:
      consumes:
      - application/json
      parameters:
      - description: Podcast payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastEpisode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Create podcast episode
      tags:
      - 'Admin: Podcasts'
  /admin/podcasts/{id}:
    delete:
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Delete podcast episode
      tags:
      - 'Admin: Podcasts'
    get:
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastEpisode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Get podcast episode
      tags:
      - 'Admin: Podcasts'
    patch:
      consumes:
      - application/json
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: string
      - description: Podcast payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastEpisode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - AdminAuth: []
      summary: Update podcast episode
      tags:
      - 'Admin: Podcasts'
...
  /podcasts:
    get:
      parameters:
      - description: Subject ID
        in: query
        name: subjectId
        type: string
      - description: Topic ID
        in: query
        name: topicId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastEpisode'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: List podcast episodes
      tags:
      - 'App: Podcasts'
  /podcasts/{id}:
    get:
      parameters:
      - description: Episode ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PodcastEpisode'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: Get podcast episode
      tags:
      - 'App: Podcasts'
  /practice/sessions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSession'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: List practice sessions
      tags:
      - 'App: Practice'
    post:
      consumes:
      - application/json
      parameters:
      - description: Session payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSessionCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSession'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: Create new practice session
      tags:
      - 'App: Practice'
  /practice/sessions/{id}:
    get:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSessionDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: Get practice session detail
      tags:
      - 'App: Practice'
  /practice/sessions/{id}/answers:
    post:
      consumes:
      - application/json
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Answer payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeAnswerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.PracticeSessionQuestion'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: Submit practice answer
      tags:
      - 'App: Practice'
...
  /revision/queue:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.RevisionItem'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      security:
      - UserAuth: []
      summary: Get revision queue
      tags:
      - 'App: Revision'
...
  /translation/do-translate:
    post:
      consumes:
      - application/json
      description: Translate a text
      operationId: do-translate
      parameters:
      - description: Set up translation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_evrone_go-clean-template_internal_controller_http_v1_request.Translate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.Translation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      summary: Translate
      tags:
      - translation
  /translation/history:
    get:
      consumes:
      - application/json
      description: Show all translation history
      operationId: history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_evrone_go-clean-template_internal_entity.TranslationHistory'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller_http_v1.ErrorResponse'
      summary: Show history
      tags:
      - translation
...
swagger: "2.0"